The func `GetServerFromRequest` should return only one
Right now it's returning a *Server and uint16. The uint16
is the ID that the specific server has. The ID is also in
the *Server as *Server.ID, so there is no need to keep the
variable ID, let alone return it. It is never once used,
and before I always put an underscore when calling this
method for the ID value. The *Server is all we really want.