Updating gulpfile
I've updated the gulpfile to be more standard. Instead of having
to run `gulp less && gulp` we now only have to run `gulp`. That's
not the only thing though. I've also added `gulp watch` and a
livereload included in the watch. I just tested it out and the
livereload server did not actually load anything new for me, which
is weird because I use this same file (a little different) at work
daily. I'll see if something is going on and if it is then I'll
update it here, but it really _should_ be good.