Updating travis.yml to use only `gulp`
Now that we've changed it from `gulp less && gulp` to `gulp` we
should also change it in our test build script. I have also switched
`npm install` to `npm update` because (although it should be a fresh
repo each time) it's better to run `update` to get the most up-to-date
versions.