Simple authentication with sessions/configuration file.
This is not the end goal, this is just something to get the web
security started. I've added two new additions to the json config
file:

```json
...
"Username": "admin",
"Password": "admin",
...
```

Obviously you should _not_ be using `admin:admin` as your user/pass
combination, but that's not up to me what other people use. The end
goal is to use _some sort_ of database (ref #45) for usernames and
passwords, and possibly allow multiple users and let each user have
their own servers/channels with their own information.

I also added some comments in other places that I did not have prior
to this commit, and removed some empty lines. The web interface now
has a "/login" page and everything gets redirected to it if a user
is not logged in and trying to access something else.

The next thing that it should do is forward 404 requests if not logged
in so it doesn't show what pages exist and what pages do not exist.

Ref #32