Renaming the session name to config.Username instead of "admin"
I was using "admin" for testing, but it really shouldn't just
be admin. I'm not 100% sure what the session name should even be.
I was thinking it should probably be a long user identification
string of some sorts, but I've yet to figure that out. I need to
do a little research on exactly _how_ sessions should be named
properly before I fix this.