GetSessionName is a func that gets the session name.
Right now it's just returning "user" as the session name. I'm not
sure if it should be doing something else to get a session name,
so I'm just returning that for now. If there's a better way to do
it then we can in the future by easily updating the func.