Create a new irc user on join if it doesn't exist.
Also create a new irc user channel related to that irc user
if it doesn't already exist. Next thing is to check if the
irc user that is JOINing is the bot itself. If it is, it should
take inventory of the users currently online, and add their
modes to Modes.

Currently the bot adds itself to irc_users and irc_user_channels
but it shouldn't. Checking to see if it is the bot itself that
is joining a channel will also be a part of fixing this.

I have also added a `TODO` in server.go for the following when
it comes to deleting a user:
  * Delete irc_users related
  * Delete irc_user_channels related