Comment out some lines. (read more...)
I haven't worked on this in months, but for some reason I have
this commented out. I need to really take a look at the whole
source and do some things to it. I'm commiting this because
I want to cherry-pick a commit into master and I'll come
back to this.