More changes, and vague commit messages. (read on)
Once this is more up-and-running and not giant commits at the
same time, there will be nice commit messages--I promise!

A lot of this commit I did a while ago, but never comitted until
now. This morning on the train I worked on the GET page for
/bots/:bot and started to create a GET page for
/bots/:bot/channel/:channel.

I also added a vim line to the top of every go file to use tabs
and not spaces becuase go fmt likes tabs!