Add `enabled` as a bool for channels.
This is needed since the lua global join and part functions are
going to be using it. Also, it's a good feature! What if we just
want to leave a channel, but only for a short amount of time, and
we want to not lose anything? We'll need an enabled channel option!